requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
playwright==1.40.0
allure-pytest==2.13.2
pytest-cov==4.1.0
requests==2.31.0
reportlab==4.0.7
PyPDF2==3.0.1
pdfplumber==0.10.3